modify authenticator, you can pass params like 'charset' etc. fix transaction bugs
