+++ docs/cn/mysql_client.md
@@ -1,7 +1,7 @@
 ```
 [mysql](https://www.mysql.com/)是著名的开源的关系型数据库，为了使用户更快捷地访问mysql并充分利用bthread的并发能力，brpc直接支持mysql协议。示例程序：[example/mysql_c++](https://github.com/brpc/brpc/tree/master/example/mysql_c++/)
 
-**注意**：brpc只支持MySQL 4.1 及之后的版本的文本协议。目前还不支持事务、Prepare statement，未来将会支持。目前支持的鉴权方式为mysql_native_password
+**注意**：只支持MySQL 4.1 及之后的版本的文本协议，支持事务，不支持Prepared statement。目前支持的鉴权方式为mysql_native_password
 
 相比使用[libmysqlclient](https://dev.mysql.com/downloads/connector/c/)(官方client)的优势有：
 
@@ -9,7 +9,7 @@
 - 支持同步、异步、半同步等访问方式，能使用[ParallelChannel等](combo_channel.md)组合访问方式。
 - 支持多种[连接方式](client.md#连接方式)。支持超时、backup request、取消、tracing、内置服务等一系列brpc提供的福利。
 - 明确的返回类型校验，如果使用了不正确的变量接受mysql的数据类型，将抛出异常。
-
+- 调用mysql标准库会阻塞框架的并发能力，使用本实现将能充分利用brpc框架的并发能力。
 # 访问mysql
 
 创建一个访问mysql的Channel：
@@ -24,7 +24,8 @@ options.protocol = brpc::PROTOCOL_MYSQL;
 options.connection_type = FLAGS_connection_type;
 options.timeout_ms = FLAGS_timeout_ms /*milliseconds*/;
 options.max_retry = FLAGS_max_retry;
-options.auth = new brpc::policy::MysqlAuthenticator("yangliming01", "123456", "test");
+options.auth = new brpc::policy::MysqlAuthenticator("yangliming01", "123456", "test", 
+    "charset=utf8&collation_connection=utf8_unicode_ci");
 if (channel.Init("127.0.0.1", 3306, &options) != 0) {
     LOG(ERROR) << "Fail to initialize channel";
     return -1;
@@ -72,22 +73,69 @@ bool Query(const butil::StringPiece& command);
 
 ​```c++
 // 返回不同类型的结果
-const MysqlReply::Auth* auth() const;
-const MysqlReply::Ok* ok() const;
-const MysqlReply::Error* error() const;
-const MysqlReply::Eof* eof() const;
+const MysqlReply::Auth& auth() const;
+const MysqlReply::Ok& ok() const;
+const MysqlReply::Error& error() const;
+const MysqlReply::Eof& eof() const;
 // 对result set结果集的操作
 // get column number
 uint64_t MysqlReply::column_number() const;
 // get one column
-const MysqlReply::Column* MysqlReply::column(const uint64_t index) const;
+const MysqlReply::Column& MysqlReply::column(const uint64_t index) const;
 // get row number
 uint64_t MysqlReply::row_number() const;
 // get one row
-const MysqlReply::Row* MysqlReply::next() const;
+const MysqlReply::Row& MysqlReply::next() const;
 // 结果集中每个字段的操作
 const MysqlReply::Field& MysqlReply::Row::field(const uint64_t index) const
 ​```
+
+事务的操作
+```c++
+ rpc::Channel channel;
+// Initialize the channel, NULL means using default options.
+brpc::ChannelOptions options;
+options.protocol = brpc::PROTOCOL_MYSQL;
+options.connection_type = FLAGS_connection_type;
+options.timeout_ms = FLAGS_timeout_ms /*milliseconds*/;
+options.connect_timeout_ms = FLAGS_connect_timeout_ms;
+options.max_retry = FLAGS_max_retry;
+options.auth = new brpc::policy::MysqlAuthenticator(
+    FLAGS_user, FLAGS_password, FLAGS_schema, FLAGS_params);
+if (channel.Init(FLAGS_server.c_str(), FLAGS_port, &options) != 0) {
+    LOG(ERROR) << "Fail to initialize channel";
+    return -1;
+}
+
+// create transaction
+brpc::MysqlTransactionOptions options;
+options.readonly = FLAGS_readonly;
+options.isolation_level = brpc::MysqlIsolationLevel(FLAGS_isolation_level);
+auto tx(brpc::NewMysqlTransaction(channel, options));
+if (tx == NULL) {
+    LOG(ERROR) << "Fail to create transaction";
+    return false;
+}
+
+brpc::MysqlRequest request(tx.get());
+if (!request.Query(*it)) {
+    LOG(ERROR) << "Fail to add command";
+    tx->rollback();
+    return false;
+}
+brpc::MysqlResponse response;
+brpc::Controller cntl;
+channel.CallMethod(NULL, &cntl, &request, &response, NULL);
+if (cntl.Failed()) {
+    LOG(ERROR) << "Fail to access mysql, " << cntl.ErrorText();
+    tx->rollback();
+    return false;
+}
+// handle response
+std::cout << response << std::endl;
+
+bool rc = tx->commit();
+```
 
 ```
