+++ src/brpc/mysql_command.h
@@ -23,38 +23,38 @@
 namespace brpc {
 // mysql command types
 enum MysqlCommandType : unsigned char {
-    COM_SLEEP,
-    COM_QUIT,
-    COM_INIT_DB,
-    COM_QUERY,
-    COM_FIELD_LIST,
-    COM_CREATE_DB,
-    COM_DROP_DB,
-    COM_REFRESH,
-    COM_SHUTDOWN,
-    COM_STATISTICS,
-    COM_PROCESS_INFO,
-    COM_CONNECT,
-    COM_PROCESS_KILL,
-    COM_DEBUG,
-    COM_PING,
-    COM_TIME,
-    COM_DELAYED_INSERT,
-    COM_CHANGE_USER,
-    COM_BINLOG_DUMP,
-    COM_TABLE_DUMP,
-    COM_CONNECT_OUT,
-    COM_REGISTER_SLAVE,
-    COM_STMT_PREPARE,
-    COM_STMT_EXECUTE,
-    COM_STMT_SEND_LONG_DATA,
-    COM_STMT_CLOSE,
-    COM_STMT_RESET,
-    COM_SET_OPTION,
-    COM_STMT_FETCH,
-    COM_DAEMON,
-    COM_BINLOG_DUMP_GTID,
-    COM_RESET_CONNECTION,
+    MYSQL_COM_SLEEP,
+    MYSQL_COM_QUIT,
+    MYSQL_COM_INIT_DB,
+    MYSQL_COM_QUERY,
+    MYSQL_COM_FIELD_LIST,
+    MYSQL_COM_CREATE_DB,
+    MYSQL_COM_DROP_DB,
+    MYSQL_COM_REFRESH,
+    MYSQL_COM_SHUTDOWN,
+    MYSQL_COM_STATISTICS,
+    MYSQL_COM_PROCESS_INFO,
+    MYSQL_COM_CONNECT,
+    MYSQL_COM_PROCESS_KILL,
+    MYSQL_COM_DEBUG,
+    MYSQL_COM_PING,
+    MYSQL_COM_TIME,
+    MYSQL_COM_DELAYED_INSERT,
+    MYSQL_COM_CHANGE_USER,
+    MYSQL_COM_BINLOG_DUMP,
+    MYSQL_COM_TABLE_DUMP,
+    MYSQL_COM_CONNECT_OUT,
+    MYSQL_COM_REGISTER_SLAVE,
+    MYSQL_COM_STMT_PREPARE,
+    MYSQL_COM_STMT_EXECUTE,
+    MYSQL_COM_STMT_SEND_LONG_DATA,
+    MYSQL_COM_STMT_CLOSE,
+    MYSQL_COM_STMT_RESET,
+    MYSQL_COM_SET_OPTION,
+    MYSQL_COM_STMT_FETCH,
+    MYSQL_COM_DAEMON,
+    MYSQL_COM_BINLOG_DUMP_GTID,
+    MYSQL_COM_RESET_CONNECTION,
 };
 
 butil::Status MysqlMakeCommand(butil::IOBuf* outbuf,
